Refactor the ingress service for istio
