+++ pkg/reconciler/knativeserving/knativeserving.go
@@ -140,8 +140,8 @@ func (r *Reconciler) transform(ctx context.Context, manifest *mf.Manifest, comp 
 		ksc.AggregationRuleTransform(manifest.Client),
 	}
 	extra = append(extra, r.extension.Transformers(instance)...)
-	extra = append(extra, ksc.IngressServiceTransform(instance))
 	extra = append(extra, ingress.Transformers(ctx, instance)...)
+	extra = append(extra, ingress.IngressServiceTransform(instance))
 	return common.Transform(ctx, manifest, instance, extra...)
 }
